% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{mvnpdf
compute the value of the density of a multivariate normal distribution}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix of n columns and p rows}

\item{mean}{a vector of means}

\item{varcovM}{a variance covariance matrix}

\item{Log}{a logical parameter to return values on a log scale, default = TRUE}
}
\value{
A list is returned with the input matrix, and a vector of density values
}
\description{
mvnpdf
compute the value of the density of a multivariate normal distribution
}
\examples{

mvnpdf(x=matrix(rep(1.96,2), nrow=2,ncol=1), Log=FALSE)

}
